CI fixes
